Add systemd config file
This change adds a systemd unit file for starting availabot.

The configuration assumes that the directory /home/ubuntu/availabot
exists and contains all files required to run the bot, including
a file called "env", from which environment variables will be set.
Currently the only variable in this file is TELEGRAM_TOKEN.

The bot is deliberately not restarted on crash. This is the easiest
way to notice when it crashes so we can investigate and fix it,
without setting up any monitoring/alerting. No SLA, no worries. :)